マイページ詳細の実装とフロントのリファクタ
