+++ client/app/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from "next/app";
 
-function MyApp({ Component, pageProps }: AppProps) {
+export function MyApp({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />;
 }
