+++ client/app/pages/messages.tsx
@@ -1,12 +1,12 @@
+import { Layout } from "../layout/Layout";
 import type { NextPage } from "next";
-import Dashboard from "../components/Top";
 import Head from "next/head";
 
 export default function Messages() {
   return (
     <>
       <title>{"messages"}</title>
-      <Dashboard title="トーク一覧" />
+      <Layout title="トーク一覧" />
     </>
   );
 }
