+++ client/app/pages/user/[id].tsx
@@ -1,4 +1,9 @@
-import UserPage from "../../components/UserPage";
+import { Layout } from "../../layout/Layout";
+import Typography from "@mui/material/Typography";
+import IconButton from "@mui/material/IconButton";
+import { Button, Card, CardActions, CardContent } from "@mui/material";
+import router from "next/router";
+import ThumbUpAltIcon from "@mui/icons-material/ThumbUpAlt";
 
 export async function getServerSideProps(ctx: any) {
   const queryId = String(ctx.query.id);
@@ -13,13 +18,38 @@ export async function getServerSideProps(ctx: any) {
   };
 }
 
-export default function GetUser(data: any) {
-  console.log(data);
-
+export default function GetUser(props: any) {
+  function Login() {
+    router.push("/users");
+  }
+  console.log(props);
   return (
-    <>
+    <div>
       <title>{"ユーザー詳細"}</title>
-      <UserPage data={data} />
-    </>
+      <Layout>
+        <IconButton onClick={Login}>
+          <>
+            いいね！
+            <ThumbUpAltIcon color={"secondary"} />
+          </>
+        </IconButton>
+        <div>
+          <Card sx={{ minWidth: 275, m: "2rem" }}>
+            <CardContent>
+              <Typography variant="h5" component="div">
+                {props.data.nickname}
+              </Typography>
+              <Typography sx={{ mb: 1.5 }} color="text.secondary">
+                {props.data.age}歳
+              </Typography>
+              <Typography sx={{ mb: 1.5 }} color="text.secondary">
+                {props.data.introduction.String}
+              </Typography>
+            </CardContent>
+            <CardActions></CardActions>
+          </Card>
+        </div>
+      </Layout>
+    </div>
   );
 }
